[refactor] create_cmd_list 리팩토링, parsing norm 수정
