+++ src/parse/tokenize/process_quotes.c
@@ -6,19 +6,11 @@
 /*   By: wonyocho <wonyocho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 17:17:02 by wonyocho          #+#    #+#             */
-/*   Updated: 2024/08/21 14:48:24 by wonyocho         ###   ########.fr       */
+/*   Updated: 2024/08/24 18:42:32 by wonyocho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-int	is_whitespace(const char c)
-{
-	if ((9 <= c && c <= 13) || c == ' ')
-		return (1);
-	else
-		return (0);
-}
 
 void	process_quotes(char *input, t_token_iter *iter)
 {
