+++ src/parse/tokenize/tokenize_util.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   util.c                                             :+:      :+:    :+:   */
+/*   tokenize_util.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: wonyocho <wonyocho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/14 15:28:16 by wonyocho          #+#    #+#             */
-/*   Updated: 2024/08/20 17:49:43 by wonyocho         ###   ########.fr       */
+/*   Updated: 2024/08/24 19:05:12 by wonyocho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,12 @@ int	check_token_type(char c)
 		return (T_DOUBLE_QUOTE);
 }
 
+void	skip_whitespace(const char *input, t_token_iter *iter)
+{
+	while (input[iter->end] && is_whitespace(input[iter->end]))
+		iter->end++;
+}
+
 int	is_quotation_str(char *str, int start, int end)
 {
 	if (str[start] == '\'' && str[end - 1] == '\'')
@@ -29,12 +35,22 @@ int	is_quotation_str(char *str, int start, int end)
 	return (0);
 }
 
-t_token	*token_lst_last(t_token *token_lst)
+int	is_redirection(const char *input, t_token_iter *iter)
 {
-	if (token_lst == NULL)
-		return (NULL);
-	while (token_lst->next != NULL)
-		token_lst = token_lst->next;
-	return (token_lst);
+	if (input[iter->end] == '<' || input[iter->end] == '>')
+	{
+		iter->end++;
+		if (input[iter->end] == input[iter->end - 1])
+			iter->end++;
+		return (1);
+	}
+	return (0);
 }
 
+int	is_whitespace(const char c)
+{
+	if ((9 <= c && c <= 13) || c == ' ')
+		return (1);
+	else
+		return (0);
+}
