+++ src/signal/signal.c
@@ -6,7 +6,7 @@
 /*   By: wonyocho <wonyocho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/06 14:08:07 by wonyocho          #+#    #+#             */
-/*   Updated: 2024/08/16 13:18:15 by wonyocho         ###   ########.fr       */
+/*   Updated: 2024/08/24 17:33:18 by wonyocho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
