+++ src/utils/error.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   error.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: chaoh <chaoh@student.42.fr>                +#+  +:+       +#+        */
+/*   By: wonyocho <wonyocho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/14 15:57:02 by wonyocho          #+#    #+#             */
-/*   Updated: 2024/08/19 20:36:33 by chaoh            ###   ########.fr       */
+/*   Updated: 2024/08/24 17:36:17 by wonyocho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
